Moving everything to do with api tests into the `tests/api` directory
