+++ test/api/runner.js
@@ -1,4 +1,4 @@
 require('env2')('config.env');
-require('./api/app.spec.js');
-require('./api/github-oAuth.spec.js');
-require('./api/github.spec.js');
+require('./specs/app.spec.js');
+require('./specs/github-oAuth.spec.js');
+require('./specs/github.spec.js');
