+++ test/api/specs/app.spec.js
@@ -1,7 +1,7 @@
 const { expect } = require('chai');
 const request = require('supertest');
 
-const app = require('../../src/api/app.js');
+const app = require('../../../src/api/app.js');
 
 describe('"/" endpoint: ', () => {
   it('should respond with a 200 and an html page', (done) => {
